requirements.txt
django
mysqlclient
BeautifulSoup4
twisted
service_identity
lxml
pytz
requests
firebase_admin
python-firebase
pyrebase
pandas
pandas_datareader
slacker
django-celery